refactor: moved tests
