+++ src/App.test.tsx
@@ -1,6 +1,6 @@
 import { render, screen } from "@testing-library/react";
 import { describe, expect, it } from "vitest";
-import App from "../App";
+import App from "./App";
 
 describe("App render", () => {
   it("renders the create idea button", () => {
